Deployment

+++ model/user.py
@@ -8,7 +8,7 @@ class User(BaseModel):
     id = PrimaryKeyField()
     insurance_id = IntegerField(null=True)
     telegram_id = IntegerField(null=True)
-    facebook_id = IntegerField(null=True)
+    facebook_id = BigIntegerField(null=True)
 
     @staticmethod
     def merge_same_user(user1: 'User', user2: 'User') -> 'User':
